Begin separating access list members from access list resources. (#30500)
* Begin separating access list members from access list resources.

Access list members are currently part of access list resources, and this can
result in access list resources that are larger than 400KB, making them
unsuitable for use by dynamo. This commit begins the process of extracting
the members from the access list resources.

The existing access list members field on the access list messages and
internal object have been left during this migration so as not to cause any
build breakages, but will be removed once the rest of the boilerplate is
in place.

* Correct spelling.

* GCI.

* Restore members in access list spec in tests/conversion to maintain compatibility during migration.