Update docs version automatically
* Update docs version automatically

Adds a new job to the post-release workflow that automatically updates
the docs version when a new release is published.

Note: we do the plugin version as well, even though it's likely that
the teleport-plugins release is still outstanding. That's okay though,
we can wait to merge the PR if we want to.

Closes #12961

* Update postrelease docs