Replace multi-line ScopedBlocks with Tabs (#30616)
* Replace multi-line ScopedBlocks with Tabs

Contributes to #30268

The documentation scope switcher tends to confuse users, and the
ScopedBlock component hides docs content based on the scope switcher. To
help remove ScopedBlocks from the docs site, this change replaces
multi-line ScopedBlocks with Tabs components if they include variations
for different scopes.

While most multi-line ScopedBlocks function like Tabs, there are a few
edge cases, which this change addresses individually.

Note that this change does not intend to remove all ScopedBlocks that
are placed inline within a paragraph. That will be the goal of a
separate change.

* Remove more ScopedBlocks (#30623)

This builds on the work done in #30616 and contributes to #30616.