Discover RDS: remove aurora engine (#30547)
'aurora' was an engine identifier for the first version of the Amazon
Aurora offering.

It reached end of life in 28th Februray 2023
https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/Aurora.MySQL56.EOL.html

We must remove it from the list of available engines, otherwise AWS will
return an error "unrecognized engine: aurora".

The backend has an equivalent PR that handles this error and queries the
remaining engines instead.