Fix ARM build process and environment setup
Fixed the ARM build process by updating the C compiler set in the Makefile and simplifying Dockerfile dependencies for the build. De-cluttered Makefile for streamlining build process. Replaced the usage of an incorrect C compiler in the Makefile and reduced the number of dependencies in Dockerfile-arm to only those necessary for building on ARM. Fixed build errors that prevented successful compilation on ARM.