Remove `Tabs` components with one `TabItem`
Backports #30769

In #30616, we replaced `ScopedBlocks` with `Tabs` components. This was
because most `ScopedBlocks` included variations on the same text for
multiple scopes, and replacing them with `TabItem`s was straightforward.
However, some replacements included only a single `TabItem`, which
displays awkwardly in the docs.

This change removes `Tabs` components with only one `TabItem`, replacing
them with body text.