Improve dependabot PR experience (#30813)
Updates the job names to reflect the language of the PR instead of
the directory name. This makes the PR titles less redundant and
makes it more obvious which language the updates are for.

Stop skipping the lint and build api workflows on dependabot PRs.
Now that the batcher isn't bundling the dependabot PRs we should
run CI checks on them prior to landing in the merge queue.