Fix max duration value when backend doesn't support it (#2034)
UI is supposed to gracefully handle the case when an error is returned from the backend when getting the maximum duration time for a request. Because of the invalid use of the Option<> object, the max duration was always set (to either 0 or the correct value) which caused issues when creating the real request.

Source-commit: gravitational/teleport.e@0e30c9ce67fec80cdc55e8dc5f059953c89c3b0f
Source-pull-requests: gravitational/teleport.e#2034