AWS OIDC: List Security Groups
When creating an EC2 Instance Connect Endpoint, we can optionally
provide a list of Security Groups to apply to it.
The default SG for the VPC is used if none SG is provided.
Sometimes the default is not what the users want.

This PR adds an endpoint that the UI will use to list possible SG for
the current VPC.
The user will then one or more SG and those are going to be applied to
the EC2 Instance Connect Endpoint