Update access duration logic and tests for dry run requests (#30876)
* "Update access duration logic and tests for dry run requests"

Adjusted the handling of 'maxDuration' for dry run requests in the access request service to prevent time drift occurrences. Max access duration in this case is now explicitly set to 7 days. Also added a test case in access_request_test.go to verify this change. The TODO note is about replacing the 'maxDuration' that currently uses point in time with 'MaxAccessDuration' that uses duration like 5 hours, 4 days etc.

* Update access duration logic and tests for dry run requests

Adjusted the handling of 'maxDuration' for dry run requests in the access request service to prevent time drift occurrences. Max access duration in this case is now explicitly set to 7 days. Also added a test case in access_request_test.go to verify this change. The TODO note is about replacing the 'maxDuration' that currently uses point in time with 'MaxAccessDuration' that uses duration like 5 hours, 4 days etc.