Release 13.3.5 (#30832)
* Release 13.3.5

* MFA Fuzzing and Json Unmarshal Panic fix

This expands the coverage of fuzzing in `lib/web/fuzz_test.go` and includes adding seeds.
From this work a new panic was discovered where a pointer is allocated then a pointer to the pointer is passed into json.Unmarshal.  It is then possible for this original pointer to remain a `nil` reference.

This pattern looks unexpected, so all cases of double pointers being passed into json.Unmarshal were changed to the more standard empty struct pointer style to avoid potential nil reference panics.

* Reword changelog entries, add security fix changelog

* Add changelog entries for new backports after rebase

---------

Co-authored-by: Mike Jensen <mike.jensen@goteleport.com>