[v12] Include consistent installation info (including Helm) across Access Request plugin docs (#30448)
* Use partial for Access Request plugin install docs

The Access Request plugin installation instructions are _almost_
standard. Make the instructions more consistent and easier to edit by
replacing them with a partial. Use our partial parameter substitution
syntax to supply the name of each plugin.

* Respond to alexfornuto feedback

* Add a missing `sudo`

Partially responds to stevenGravy feedback

* Add consistent Helm details

Ensure all Access Request plugin guides include the same Helm-related
steps. Excluding the Jira plugin because that guide is out of date and
will be addressed in a separate change.

Also removes the "Feedback" section from each Access Request plugin
guide.  These are out of step with the rest of the docs, which don't
include a feedback section, and do not reflect the current location of
the Access Request plugin source.

* Spelling and linter fixes