[v13] AWS OIDC: Configure IAM for EC2 Instance Connect Endpoint (#30948)
* AWS OIDC: Configure IAM for EC2 Instance Connect Endpoint

This PR adds a new teleport command that sets up the required
permissions to use EC2 Instance Connect Endpoint to connect to an EC2
instance.

It also adds a oneoff script that runs this command.
The goal is to give a user a script for them to run, the script then
downloads teleport and calls AWS APIs to create the inline policy.

* add const and lowercase error

* use correct client