build: Build webassets for CentOS 7 with buildbox-connect (#30322)
* build: Build webassets for CentOS 7 with Ubuntu buildbox

Build the webassets for the CentOS 7 binaries using the standard Ubuntu
buildbox prior to building the binaries. This allows us to use a later
version of node.js to build the UI than what is available on CentOS 7.

The `release-unix` and `clean` targets need to be rejigged a little as
`release-unix` first cleans any pre-build webassets. So we add a new
target `release-unix-preserving-webassets` that only cleans the build
and not the webassets and update the Makefile in `build.assets` to call
this new target when building CentOS 7 binaries.

* build: Conditionally build enterprise webassets

Only build enterprise webassets if the enterprise submodule is present.
For a normal release, this will always be the case, but for local
testing and people wishing to build just the OSS release, the enterprise
submodule may not be present.

* build: Build webassets with buildbox-connect

Use the Teleport Connect buildbox for building the webassets. That
ensures the same version of node is used for building all the node
stuff; the web UI as well as Teleport Connect. It makes maintenance of
the buildboxes easier as we can move to using only the one buildbox for
node in the future.

* build: Pre-build webassets for non-centos7 build targets

Add a dependency on `webassets` for the non-centos7 fips build/release
targets in build.assets/Makefile. These targets were recently changed to
actually use centos7 anyway, so we do need to pre-build webassets as
centos7 will soon no longer be able to build them.

NOTE: These updated targets should probably be removed. They are
effectively duplicates of their centos7 counterparts now and should
probably just chain to those targets. The `release-fips` target is
actually broken at the moment as it needs to use `scl enable ...` to
build now it uses centos7.

* Replace "standard Ubuntu buildbox" in comments with buildbox-connect

---------

Co-authored-by: Grzegorz Zdunek <grzegorz.zdunek@goteleport.com>
Co-authored-by: Grzegorz Zdunek <gzdunek@users.noreply.github.com>