Apply misc UI improvements to Connect My Computer (#31028)
* Increase wait time when switching from setup to status

It was a little bit too short. Preferably, there should be an additional
success animation of some kind at the end, but we're going to leave this
out for now.

* Adjust copy of the downloading state

"Verifying" sounds like we're doing some actual verification with shasum
or something, but we're basically just checking the version of the
binary.

* Make the connect/disconnect button larger

It has 100% width, but it was a little bit too short for such a wide
button.

* Status doc: Add unmountOnExit to Transition

This way the child component will get unmounted on the exit state.
Otherwise we'd attempt to render it even if there was no agent node.

* Change text casing in nav menu of Connect My Computer

We recently started to apply title case to all labels, we'll probably
have to start doing to for buttons as well.