ci: Use "post-release" environment in update-docs post-release workflow (#30937)
Add an `environment: post-release` clause to the `update-docs-version`
job in the `post-release.yaml` workflow so it can access the `APP_ID`
var and the `PRIVATE_KEY` secret. This allows it to generate a github
token with the permissions of the installed application referenced by
`APP_ID`.

App: https://github.com/organizations/gravitational/settings/apps/teleport-post-release-automation