Relax CA Rotation test assertions (#31098)
Changes the assertions to only care that an error has occurred, and
not that the error is a bad certificate error from the tls handshake.
When the tests fail the error is a bad pipe error which seems to
be caused by the handshake failing prior to the grpc settings being
written. In either scenario the connection is terminated due to
incorrect certificates so asserting that an error was received should
be satisfactory.