Ensure access list data integrity. (#31044)
This commit adds locking to the access list service. The concern here is that,
without this locking and the additional logic here, it's possible to add
members to access lists that don't exist. Furthermore, if a user then creates
an access list with the same name as the access list being used for creating
these members, then the access list will have members added to it.

The locks and logic here should ensure that members are only added to lists
that exist and that users are properly cleaned up after removing a list
without the possibility of orphaned members.

Co-authored-by: rosstimothy <39066650+rosstimothy@users.noreply.github.com>