Move the `tsh` config file guidance (#30953)
Backports #30437

Fixes #27817

Move our guidance on the `tsh` config files to the `tsh` guide in the
"Connect your Client" section and out of the CLI reference. This is
because the CLI reference focuses on commands, while the general `tsh`
guide covers a number of topics, making it a more appropriate place for
the config file guide.