AWS OIDC - List RDS: add VPC ID (#30971)
* AWS OIDC - List RDS: add VPC ID

When deploying a Teleport service using the AWS OIDC DeployService, we
can provide the SecurityGroups to be applied to the network
configuration of the service.

Using the default SG for the VPC (vpc is inferred from the used subnets)
might not always be the best one (it is usually locked down).

The VPC is required to list the possible SecurityGroups.

Currently, there's no VPC information in the database.

This PR adds a new field to the AWS.RDS fields to indicate which VPC is
this Database in.

When running the DeployService, the user should be able to list the
Security Groups and then pick one or many to apply to the service.

* vpcId to vpcID