Fix the access list lockName in the backend service. (#31279)
In the backend service, using slashes in the lock name seems to make the
sqlite backend (and possibly others) error with the following message:

```
special characters are not allowed in resource names, please use name
composed only from characters, hyphens, dots, and plus signs:
".locks//access_list/test-access-list"
```

It looks like this is due to the fact that the `backend.Key` that was
previously being used was inserting a forward slash in the front of the
string for generating the lock name. It has been replaced by `strings.Join`,
which will not result in a string that has a prepended forward slash.