Fix AuthOrIsOwner test. (#2103)
The AuthOrIsOwnerTest was running into issues with the newly added locks in
the access list backend because the clock being used is a FakeClock and
doesn't advance, so lock contention is never retried. We get around this by
not parallelizing this, ensuring that lock contention is handled by the end
of a test case.

Source-commit: gravitational/teleport.e@4f99aa6466679e6581d2635ee03cc4c01eb76343
Source-pull-requests: gravitational/teleport.e#2103