AWS OIDC EICE: fix connection set up (#31209)
* AWS OIDC EICE: fix connection set up

Sending an SSH Public Key and creating the Tunnel are done in two
distinct moments.

When deciding wether to use a tunnel or a direct connection, we are now
creating the Tunnel and return the net.Conn which connects directly to
the EC2 Private IP.

The SSH Public Key is sent afterwards when the SSH protocol starts.

It also adds a metrics to the connection.

* improve error message