AWS OIDC - List EC2: add instance id as label
When listing EC2 instances we should be able to tell which ones were
already added.
To do so, we should use the instance-id which is unique.
We can't use the Spec.CloudMetadata.AWS.InstanceID because predicate
can't use fields that are optional.

This PR adds the InstanceID as label.
It uses `teleport.dev/instance-id` label key.

When creating a Node from WebUI, it will send the same set of labels
that was received.
So, WebUI will send this same label.

When listing EC2 instances, WebUI queries the backend for a list of
Nodes of type EC2 EICE and the same `instance-id` and uses the result to
show which ones were already added.