Slow down API calls and make sync duration configurable. (#2077)
Okta seems to have fairly aggressive rate limits in place. The most noteworthy
ones:

* Apps seem to default to 100 per minute, and we need to read each app.
* There is a 40 API call per 10 second universal rate limit, which we can
  easily hit when using 10 API calls per second.

To address this, API calls have been slowed down to 4 per second, yielding a
max of 240 per minute. Additionally, the default time between synchronizations
is now 10 minutes.

For 500 apps and 10000 groups, that should yield:
* 3 calls to paginate through apps.
* 500 calls to retrieve each app.
* 1 call for listing all of the groups.

This should result in 504 calls, taking roughly 126 seconds. This would be
a fairly large Okta setup by my observation, so I think this should handle
things pretty well with the downside of having a slower synchronization time.
According to Okta, the average number of applications in a large company is
211 (https://www.okta.com/au/businesses-at-work/). The above calculation would
become, assuming the same number of groups:

* 2 calls to paginate through apps.
* 211 calls to retrieve each app.
* 1 call for listing all of the groups.

53.5 seconds, which is a much safer number.

If users are finding the synchronization time too slow, the duration between
synchronization calls is now configurable, which should allow users to have
a more responsive sync service.

I think limiting the API calls to 4 per second is unfortunately necessary. I
don't suspect this will cause an issue with assignments, as we check these
every 5 minutes. This *could* potentially double the synchronization time if
every single application is checked as part of the assignment reconciler. It's
unlikely that this will be the case, however, and should only increase things
by a few seconds.

Co-authored-by: Roman Tkachenko <roman@goteleport.com>

Source-commit: gravitational/teleport.e@cc011c443e3d5d8fd23e669c940a9ab92f781cd5
Source-pull-requests: gravitational/teleport.e#2077