AWS OIDC - DeployService: add optional Security Groups (#31268)
When deploying a Teleport service using the AWS OIDC DeployService, we
can provide the SecurityGroups to be applied to the network
configuration of the service.

Using the default SG for the VPC (vpc is inferred from the used subnets)
might not always be the best one (it is usually locked down).

This PR adds an optional new field that clients can use to be explicit
about the SGs they want.

We also added a new permission when configuring the IAM Permissions for
the DeployService, to allow the usage of the ListSecurityGroups action.
Users can see the list of SG and then pick which ones they want.