AWS OIDC - List SecurityGroups: add Inbound and Outbound Rules
This PR adds the SecurityGroups Inbound and Outbound Rules to the
Teleport representation of SecurityGroups.

This will allow the user configuring the SGs for DeployService and EICE
access to quickly understand what rules are available for each
SecurityGroup.
This way, they can pick the ones that match their security requirements.