Edit Server Access intro guide architecture info (#31494)
Fixes #23145

Clarify how to run the Teleport SSH Service in a Kubernetes environment
by editing the architecture information in the server access "Getting
Started" guide. This is an appropriate place for this information, since
this is the guide that provides instructions on protecting a server with
Teleport.

To make this clarification, mention the Teleport SSH Service and Proxy
Service explicitly, explaining how they interact. Then explain how this
setup works in a Kubernetes environment.