Make the "manual" backport tool behave more like the bot
When backports are opened with `make backport`, the resulting
PR has very little information about the actual change. The
title and body just say "Backport XYZ".

Update the tool to open PRs with a title that is the same as the
original PR (with the correct version prefix). Also ensure that
the body of the original PR is preserved, along with a reference
to the original PR so it's easy to track that this was a backport.