Make sure Teleport sessions use the user login state. (#31363) (#31614)
The user login state is being generated properly but not being used for
the various sessions. UserLoginState is now properly integrated into
sessions. Additionally, it was noticed that user login state was not being
used for Github logins, so it's been added there as well.

Additionally, getUserOrLoginState has been exposed for use by the
enterprise repo.

Co-authored-by: Lisa Kim <lisa@goteleport.com>