Remove unsafe usage of resource id
The transport used by app service was using a sync.Map keyed by
the resource id to store app servers. However, the id of a resource
is not guaranteed to be unique which could cause apps to be missed.
The sync.Map was removed in favor of a mutex and iterating over the
configured servers slice directly.