React to version updates faster (#31652)
The version shown in the UI's Help & Support page is populated by
querying for all auth servers in the cluster and taking the first
one. It's possible that after an upgrade the old auth server hasn't
expired, so we may incorrectly report the old version for several
minutes.

Fix this by:
- using the version from the newest auth server
- using the embedded version for the proxy version