Deflake TestTeleportProcessAuthVersionCheck (#31688)
This test uses a broken approach to try and allocate a local port
for the auth service. It starts a listener without specifying a port,
grabs the port that the OS allocates, and then closes the listener
and assumes that the port that was free at the time of this check will
remain free by the time the test needs to open a 2nd listener.

Instead, start the auth service first, allow the OS to allocate a free
port, and *use that port* in the SSH service's config. This ensures
that the port remains free since we keep the listener open the whole
time.

Fixes #19752