Improve logging for the upload completer (#31569)
We've seen occaisional OOMs for large multipart uploads.
This change improves the logging to help with troubleshooting and
also catches errors early by refusing to submit API requests with
too many parts.

(Note: this should resolve the OOM, which happens in the AWS SDK,
but does not solve the underlying cause)