Remove bloat check (#2188)
The check cannot be run because there is no way to determine the
branch to use to checkout oss from the push event. We would have
to run the check on merged PRs to get the branch information. There
is likely enough signal by just running the check in OSS - so the
check is removed from e instead.

Source-commit: gravitational/teleport.e@94825bec6402192f3d0bddc2f55ab8294f47250f
Source-pull-requests: gravitational/teleport.e#2188