Validate desktop names (#31767)
Desktops cannot contain '.' characters in their names, as we use SNI
to pass the desktop name in the TLS handshake. This wasn't a problem
in the past when Teleport was the only thing that created desktops,
as we were sure to sanitize names. With non-AD access we encourage
users to register their own desktops via API or tctl, and they often
encounter a confusing TLS error when they use an invalid name.

With this change:

    ➜ tctl create -f desktop-resource.yaml
    ERROR: invalid name "this-is-a.test": desktop names cannot contain periods

Fixes #31644