Add docs to Kubernetes per-Resource RBAC (#31483)
* Add docs to Kubernetes per-Resource RBAC

This PR adds documentation for the Kubernetes per-Resource RBAC and
supported verbs introduced in Teleport 14.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* initial review

* fix lint

* Apply suggestions from code review

Co-authored-by: Paul Gottschling <paul.gottschling@goteleport.com>

* fix bad reference

* Restructure the Resource Requests guide (#31637)

Move the subsection on requesting access to Kubernetes resources to an
H3. This way, we can avoid an unslightly H6 heading. Also, since this
section relates to requesting access, not restricting access, it's more
discoverable in a separate H3.

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>
Co-authored-by: Paul Gottschling <paul.gottschling@goteleport.com>