[v14] AWS OIDC: Set up integration with a single command (#31791)
* AWS OIDC: Set up integration with a single command

Creating an AWS OIDC Integration requires a lot of clicks, copy/paste,
navigation between tabs.

This PR adds a single teleport command that creates all the required
resources in AWS:
- AWS OIDC Identity Provider that uses Teleport as source
- AWS IAM Role that can be used by this Identity Provider

This role, will then have inline policies allowing multiple features in
Teleport.
Example: a policy must be added to this role that has
`rds:DescribeInstances` and `rds:DescribeClusters` so that the user can
use the integration to enroll RDS Databases.

* improve docs

* improve godoc