Fix inconsistency with `kubectl auth can-i get pods/<pod> -n <ns>` (#31875)
This PR fixes an inconsistency with our own `SelfSubjectAccessReview`
where if the user sends a non-empty resource name `<pod>` and the user
holds a `kubernetes_resources` with `kind: namespace`, we ended up
validating the resource name <pod> against the namespace name in the
`kubernetes_resource`. This behavior is incorrect if the resource is a
namespaceScoped resource, we must skip the validation of it because
namespace grants access to all resources within itself.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>