[v14] Fix directory sharing for non-ascii directory names (#31925)
* Fix directory sharing for non-ascii directory names

Previously we were sending the shared directory's name for PreferredDosName
(truncated at 7 characters per the spec). This works in many cases, but
if the directory name contains non-ASCII characters it could lead to a
panic when we tried to slice the name in the middle of a non-ascii UTF
character.

Instead, we now send a fixed string of "FILE" for PreferredDosName when
sharing a filesystem (directory). This value is effectively ignored, the
actual drive name is always taken from the DeviceData field which can handle
non-ASCII characters.

* adds note re: multiple drives in the future

* fixes test_handle_client_id_confirm

* add test_device_announce_header_encode_with_non_ascii